Restructure lib/utils/game
