+++ lib/game/index.js
@@ -1,0 +1,15 @@
+const model = require('./model')
+const {
+  getGameState,
+  getEntity,
+  getAllEntities,
+  isFieldEmpty
+} = require('./utils')
+
+module.exports = {
+  model,
+  getGameState,
+  getEntity,
+  getAllEntities,
+  isFieldEmpty
+}
