+++ lib/game/model.js
@@ -1,6 +1,39 @@
 const _ = require('lodash')
-const update = require('immutability-helper')
-const uid = require('uid')
+const api = require('./api')
+const {
+  getGameState,
+  updateGameState,
+  createEntity,
+  getEntity,
+  getAllEntities
+} = require('./utils')
+
+module.exports = ({ state, components, reducers, effects }) => {
+  // TODO: enforce that there are no components which have effects and reducers with the same name
+
+  // we need to ensure that every component handles the update action which is triggered after every game step
+  // if a component has no update handler _.noop is added as effect
+  const componentsWithUpdateHandler = getComponentsWithUpdateHandler(components)
+
+  const gameAPI = api(_.keys(components))
+
+  return {
+    namespace: 'game',
+    state: init(state),
+    reducers: _.assign(
+      {},
+      getComponentsReducers(componentsWithUpdateHandler),
+      gameAPI.reducers,
+      getGlobalReducers(reducers)
+    ),
+    effects: _.assign(
+      {},
+      getComponentsEffects(componentsWithUpdateHandler),
+      gameAPI.effects,
+      getGlobalEffects(effects)
+    )
+  }
+}
 
 function init (state) {
   const current = _.assign(
@@ -20,131 +53,6 @@ function init (state) {
     prev: null,
     current,
     next: null
-  }
-}
-
-function getGameState (state) {
-  if (!state.isGameStepCompleted) {
-    return state.next
-  }
-
-  return state.current
-}
-
-function updateGameState (state, changes) {
-  if (!state.isGameStepCompleted) {
-    return update(state, { next: changes })
-  }
-
-  return update(state, { current: changes })
-}
-
-function createEntity (entity) {
-  if (!entity.id) {
-    entity.id = uid()
-  }
-
-  // TODO: enforce component requirements
-
-  return entity
-}
-
-function getEntity (id, game) {
-  return game.entities[id]
-}
-
-function isFieldEmpty (game, x, y) {
-  return !(game.tiles[y] === undefined || game.tiles[y][x] === undefined || game.tiles[y][x] === 0)
-}
-
-function deleteEntity (id, game) {
-  return _.omitBy(game.entities, (entitiy) => {
-    return entitiy.id === id
-  })
-}
-
-function getAllEntities (componentType, game) {
-  return _.filter(game.entities, (entity) => !_.isUndefined(entity[componentType]))
-}
-
-function getGameAPI (componentNames) {
-  return {
-    reducers: {
-      createEntity: (state, { data }) => {
-        const entity = createEntity(data)
-
-        return updateGameState(state, {
-          entities: {
-            [entity.id]: { $set: entity }
-          }
-        })
-      },
-
-      deleteEntity: (state, { data }) => {
-        return updateGameState(state, {
-          entities: {
-            $set: deleteEntity(data.id, state)
-          }
-        })
-      },
-
-      beginStep: (state) => {
-        return update(state, {
-          next: { $set: state.current },
-          isGameStepCompleted: { $set: false }
-        })
-      },
-
-      _completeStep: (state) => {
-        if (state.isGameStepCompleted) {
-          return state
-        }
-
-        return update(state, {
-          prev: { $set: state.current },
-          current: { $set: state.next },
-          next: { $set: null },
-          isGameStepCompleted: { $set: true }
-        })
-      }
-    },
-
-    effects: {
-      completeStep: (state, data, send) => {
-        _.forEach(componentNames, (name) => {
-          send(`game:${name}.update`, {}, _.noop)
-        })
-
-        send('game:_completeStep', {}, _.noop)
-      }
-    }
-  }
-}
-
-function game ({ state, components, reducers, effects }) {
-  // TODO: enforce that there are no components which have effects and reducers with the same name
-
-  // we need to ensure that every component handles the update action which is triggered after every game step
-  // if a component has no update handler _.noop is added as effect
-  const componentsWithUpdateHandler = getComponentsWithUpdateHandler(components)
-
-  const gameAPI = getGameAPI(_.keys(components))
-
-  return {
-    namespace: 'game',
-    state: init(state),
-    reducers: _.assign(
-      {},
-      getComponentsReducers(componentsWithUpdateHandler),
-      gameAPI.reducers,
-      getGlobalReducers(reducers)
-    ),
-    effects: _.assign(
-      {},
-      getComponentsEffects(componentsWithUpdateHandler),
-      gameAPI.effects,
-      getGlobalEffects(effects)
-    )
   }
 }
 
@@ -270,12 +178,4 @@ function getGlobalEffects (effects) {
 
     return handlers
   }, {})
-}
-
-module.exports = {
-  engine: game,
-  getGameState,
-  getEntity,
-  getAllEntities,
-  isFieldEmpty
 }
